LayoutTests/imported/w3c/web-platform-tests/html/webappapis/microtask-queuing/queue-microtask-cross-realm-callback-report-exception-expected.txt

CONSOLE MESSAGE: Error: PASS


PASS queueMicrotask() reports the exception from its callback in the callback's global object
